feat: Login with phoneNumber and password
-Refactor the codes

[Finishes
