+++ src/helpers/registerFunction.helper.js
@@ -1,5 +1,5 @@
 import handleSignup from './handleSignup.help';
-import obj from '../screens/welcome/WelcomeScreen';
+import obj from '../database/connectToDb';
 
 const { myDb } = obj;
