+++ src/screens/welcome/WelcomeScreen.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { View, Text, Image, TouchableOpacity } from 'react-native';
 import PropTypes from 'prop-types';
-import { openDatabase } from 'react-native-sqlite-storage';
 import welcome from '../../helpers/welcome.helper';
 import styles from '../../styles/allStyles';
 import checkNetworkConnectivity from '../../helpers/checkInternetConnection.helper';
@@ -10,7 +9,6 @@ const phoneHistoryStoryStoreLogo = require('../../../assets/img/phone-history-st
 const myLoadingIcon = require('../../../assets/img/loading-icon.gif');
 
 const { allStyles } = styles;
-const myDb = openDatabase({ name: 'nezadigitalstore.db', createFromLocation: 1 });
 
 /**
  * @class
@@ -24,17 +22,6 @@ class WelcomeScreen extends Component {
   constructor() {
     super();
     this.state = {};
-    myDb.transaction((txn) => {
-      txn.executeSql(
-        "SELECT name FROM sqlite_master WHERE type='table' AND name='users'", [],
-        (tx, res) => {
-          if (res.rows.length === 0) {
-            txn.executeSql('DROP TABLE IF EXISTS users', []);
-            txn.executeSql('CREATE TABLE IF NOT EXISTS users(id INTEGER PRIMARY KEY, phoneNumber VARCHAR(20), firstName VARCHAR(255), lastName VARCHAR(255), email VARCHAR(255), "password" VARCHAR(255), age INTEGER, isRegistered BOOLEAN DEFAULT false);', []);
-          }
-        },
-      );
-    });
   }
 
   componentDidMount = () => {
@@ -77,4 +64,4 @@ WelcomeScreen.propTypes = {
   navigation: PropTypes.object.isRequired,
 };
 
-export default { WelcomeScreen, myDb };
+export default WelcomeScreen;
